# openapi.yaml
openapi: 3.0.0
info:
  title: Enterprise Authentication Service API
  description: This service handles user authentication and authorization for the enterprise application.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/auth
    description: Local development server
paths:
  /login:
    post:
      summary: Authenticate a user
      description: |
        Authenticates a user with an email and password. A successful login returns an access token and sets a secure HttpOnly refresh token cookie.
      parameters:
        - name: X-Device-Id
          in: header
          description: Unique identifier for the user's device.
          required: true
      schema:
          type: string
          example: "abcdef1234567890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: victoroluborode@example.com
                  description: The user's email address.
                password:
                  type: string
                  example: securepassword123
                  description: The user's password.
      responses:
        '200':
          description: Login successful. Returns an access token and sets a refresh token cookie.
          headers:
            Set-Cookie:
              description: The refresh token is set as a secure, HttpOnly cookie.
              schema:
                type: string
                example: "refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ2aWN0b3JvbHVib3JvZGUiLCJpYXQiOjE2ODcyMDA2NDgsImV4cCI6MTY4NzIwNDI0OH0.SFLy7gU5T1hS4fR5jK8y8f8JvL2fQv4oD3C8gL8q; HttpOnly; Secure"
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ2aWN0b3JvbHVib3JvZGUiLCJpYXQiOjE2ODcyMDA2NDgsImV4cCI6MTY4NzIwNDI0OH0.SFLy7gU5T1hS4fR5jK8y8f8JvL2fQv4oD3C8gL8q"
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /register:
    post:
      summary: Register a new user
      description:  |
       Creates a new user account and sends a verification email.
      parameters:
        - name: X-Device-Id
          in: header
          description: Unique identifier for the user's device.
          required: true
          schema:
            type: string
            example: "abcdef1234567890"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - fullname
              properties:
                email:
                  type: string
                  example: newuser@example.com
                password:
                  type: string
                  example: StrongPassword123!
                fullname:
                  type: string
                  example: New User
      responses:
        '201':
          description: User created successfully.
          headers:
            Set-Cookie:
              description: The refresh token is set as a secure, HttpOnly cookie.
              schema:
                type: string
                example: "refreshToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ2aWN0b3JvbHVib3JvZGUiLCJpYXQiOjE2ODcyMDA2NDgsImV4cCI6MTY4NzIwNDI0OH0.SFLy7gU5T1hS4fR5jK8y8f8JvL2fQv4oD3C8gL8q; HttpOnly; Secure"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully."
                  accessToken:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ2aWN0b3JvbHVib3JvZGUiLCJpYXQiOjE2ODcyMDA2NDgsImV4cCI6MTY4NzIwNDI0OH0.SFLy7gU5T1hS4fR5jK8y8f8JvL2fQv4oD3C8gL8q"
                  user:
                    $ref: '#/components/schemas/UserResponse'
                  roles:
                    type: array
                    items:
                      type: string
                      example: "user"
        '400':
          description: User already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /verify-email:
    get:
      summary: Verify a user's email address
      description: Verifies a user's email using a token and token ID from a verification link.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          example: "abcde"
        - name: tokenId
          in: query
          required: true
          schema:
            type: string
          example: "12345"
      responses:
        '200':
          description: Email verified successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email successfully verified"
        '401':
          description: Invalid or expired token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /resend-verification-email:
    post:
      summary: Resend verification email
      description: Sends a new email verification link to a user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  example: user@example.com
      responses:
        '200':
          description: Verification email resent successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Verification email resent successfully"
                  verificationlink:
                    type: string
                    example: "http://localhost:3000/api/auth/verify-email?token=abcde&tokenId=12345"
        '400':
          description: Email already verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /post:
    post:
      summary: Create a new post
      description: |
        Creates a new post. This route requires a valid access token and the 'post:create' permission.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
              properties:
                title:
                  type: string
                  example: "My First Blog Post"
                content:
                  type: string
                  example: "This is the content of my first post."
      responses:
        '201':
          description: Post created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized due to invalid token or missing permissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /change-password:
    post:
      summary: Change user password
      description: Allows a logged-in user to change their password.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentpassword
                - newpassword
              properties:
                currentpassword:
                  type: string
                  example: "securepassword123"
                newpassword:
                  type: string
                  example: "NewStrongPassword456"
      responses:
        '200':
          description: Password changed successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password changed successfully"
        '401':
          description: Invalid current password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /forgot-password:
    post:
      summary: Request a password reset link
      description: Sends a password reset link to the user's email address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  example: user@example.com
      responses:
        '200':
          description: |
            A password reset link has been sent if an account with that email exists.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If an account with that email exists, a password reset link has been sent."
                  resetPasswordLink:
                    type: string
                    example: "http://localhost:3000/api/auth/reset-password?tokenId=abcde&token=12345"
  /reset-password:
    post:
      summary: Reset password
      description: Resets the user's password using a valid token and token ID from the reset link.
      parameters:
        - name: tokenId
          in: query
          required: true
          schema:
            type: string
          example: "abcde"
        - name: token
          in: query
          required: true
          schema:
            type: string
          example: "12345"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newpassword
              properties:
                newpassword:
                  type: string
                  example: "NewStrongPassword456"
      responses:
        '200':
          description: Password reset successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successful"
        '401':
          description: Invalid or expired token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts:
    get:
      summary: Retrieve all posts
      description: |
        Returns a list of all posts. This route requires a valid access token and the 'post:read' permission.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Posts retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Access Granted"
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostWithUser'
        '401':
          description: Unauthorized due to invalid token or missing permissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
          example: "65e23c7c2518e3a2410a0e70"
        email:
          type: string
          example: "user@example.com"
        fullname:
          type: string
          example: "User Name"
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid credentials"
    Post:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "My First Blog Post"
        content:
          type: string
          example: "This is the content of my first post."
        userId:
          type: string
          example: "65e23c7c2518e3a2410a0e70"
        createdAt:
          type: string
          format: date-time
          example: "2025-08-13T10:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-08-13T10:00:00.000Z"
    PostWithUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "My First Blog Post"
        content:
          type: string
          example: "This is the content of my first post."
        userId:
          type: string
          example: "65e23c7c2518e3a2410a0e70"
        createdAt:
          type: string
          format: date-time
          example: "2025-08-13T10:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-08-13T10:00:00.000Z"
        user:
          type: object
          properties:
            fullname:
              type: string
              example: "Victor Oluborode"
            email:
              type: string
              example: "victor@example.com"
...